sprites: input_flag seems to try some kind of column interlacing.
col_interlaced_paths point to this data (each nibble is actually coded
on a full byte, but high nibble is always zero):

    0 8 4 C
    2 A 6 E
    1 9 5 D
    3 B 7 F

...which corresponds to some kind of columns paths in:

    0 1 2 3
    4 5 6 7
    8 9 A B
    C D E F

When enabling input_flag, output picture looks like a mess. I don't know
what this flag really is for, and if it is correctly implemented in the
rest of the current pokanalysis decoder.