Rewrite project in Rust, with a (static) web frontend
The project was hardly buildable anymore due to CPython 2 and GTK2
dependencies. This was also an opportunity for me to learn the Rust
language. The web view was picked for portability and accessibility
reasons, instead of requiring users to build the project.

The frontend (web UI) and backend (rust extractor) are now well split:
the extractor dumps images and a JSON database, and the web UI renders
that statically.

Compared to the C version, the Rust code contains many improvements with
regards to fixes and features. A few things are not yet included though:

1. the disasm, but that wasn't usable in any way
2. the wild grass and water Pokémon are exported in the database but not
   yet exposed to the map

Note that overall the code still need a lot of improvements in style and
organisation, in particular because this is my first Rust project. The
dependencies are kept to a minimum: png for exporting images, and serde
for serializing the JSON database.

The frontend is barely functional because I'm also not a web expert. I
also avoided any dependency so this won't be a maintenance burden in the
future; raw HTML/CSS/Javascript should be working 10 years from now,
hopefully.